+++ src/method_wrappers/wrapper_transferentropy_regular.jl
@@ -200,9 +200,8 @@ end
         response::AbstractArray{<:Real, 1}, 
         k::Int, l::Int, m::Int; 
         η = 1, τ = 1, 
-        estimator = VisitationFrequency(), 
-        n_subdivs = 1,
-        b = 2)
+        estimator = VisitationFrequency(b = 2), 
+        n_subdivs = 1)
 
 ## TE estimation with default discretization scheme(s)
 
@@ -229,7 +228,7 @@ all embedding components. `η` is the prediction lag.
     (this way, TE is computed over two separate partitions). Unless `n_subdivs = 0`,
     make sure that `n_subdivs` is the same across analyses if they 
     are to be compared. T TE 
-- **`b`**: Base of the logarithm. The default (`b = 2`) gives the TE in bits.
+
 ## More about the embedding
 
 To compute transfer entropy, we need an appropriate delay embedding 
@@ -286,9 +285,8 @@ function te_reg(source::AbstractArray{<:Real, 1},
         response::AbstractArray{<:Real, 1}, 
         k::Int, l::Int, m::Int; 
         η = 1, τ = 1, 
-        estimator = VisitationFrequency(), 
-        n_subdivs = 2,
-        b = 2)
+        estimator = VisitationFrequency(b = 2), 
+        n_subdivs = 2)
 
     k + l + m >= 3 || throw(ArgumentError("`dim = k + l + m` must be 3 or higher for regular TE"))
 
@@ -307,7 +305,7 @@ function te_reg(source::AbstractArray{<:Real, 1},
     binning_scheme = map(n-> RectangularBinning(n), nbins_coarsest:nbins_finest)
     # Compute TE over different partitions
     # ====================================
-    tes = map(binscheme -> transferentropy(pts, vars, binscheme, estimator, b = b), binning_scheme)
+    tes = map(binscheme -> transferentropy(pts, vars, binscheme, estimator), binning_scheme)
 
     return tes
 end
@@ -320,8 +318,7 @@ end
         k::Int, l::Int, m::Int,
         binning_scheme::Vector{RectangularBinning}; 
         η = 1, τ = 1, 
-        estimator = VisitationFrequency(), 
-        b = 2)
+        estimator = VisitationFrequency(b = 2))
 
 # TE with user-provided discretization scheme(s)
 
@@ -348,7 +345,6 @@ using an embedding delay of `τ` across all embedding components.
 - **`η`**: The prediction lag. Default is `η = 1`.
 - **`estimator`**: The transfer entropy estimator to use. The default 
     is `VisitationFrequency()`.
-- **`b`**: Base of the logarithm. The default (`b = 2`) gives the TE in bits.
 
 ## More about the embedding
 
@@ -398,8 +394,7 @@ function te_reg(source::AbstractArray{<:Real, 1},
         k::Int, l::Int, m::Int,
         binning_scheme::Vector{RectangularBinning},
         η = 1, τ = 1, 
-        estimator::TransferEntropyEstimator = VisitationFrequency(), 
-        b = 2)
+        estimator::TransferEntropyEstimator = VisitationFrequency(b = 2))
 
     k + l + m >= 3 || throw(ArgumentError("`dim = k + l + m` must be 3 or higher for regular TE"))
 
@@ -407,7 +402,7 @@ function te_reg(source::AbstractArray{<:Real, 1},
 
     # Compute TE over the partitions constructed from the provided binning schemes
     # ====================================
-    tes = [transferentropy(pts, vars, bs, estimator, b = b) for bs in binning_scheme]
+    tes = [transferentropy(pts, vars, bs, estimator) for bs in binning_scheme]
 
     return tes
 end
@@ -417,8 +412,7 @@ function te_reg(source::AbstractArray{<:Real, 1},
         k::Int, l::Int, m::Int,
         binning_scheme::RectangularBinning,
         η = 1, τ = 1, 
-        estimator::TransferEntropyEstimator = VisitationFrequency(), 
-        b = 2)
+        estimator::TransferEntropyEstimator = VisitationFrequency(b = 2))
 
     k + l + m >= 3 || throw(ArgumentError("`dim = k + l + m` must be 3 or higher for regular TE"))
 
@@ -426,7 +420,7 @@ function te_reg(source::AbstractArray{<:Real, 1},
 
     # Compute TE over the partitions constructed from the provided binning schemes
     # ====================================
-    te = transferentropy(pts, vars, binning_scheme, estimator, b = b)
+    te = transferentropy(pts, vars, binning_scheme, estimator)
 
     return te
 end